update curl: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@53610 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/curl/tests/unit/unit1323.c
@@ -5,7 +5,7 @@
  *                            | (__| |_| |  _ <| |___
  *                             \___|\___/|_| \_\_____|
  *
- * Copyright (C) 1998 - 2020, Daniel Stenberg, <daniel@haxx.se>, et al.
+ * Copyright (C) Daniel Stenberg, <daniel@haxx.se>, et al.
  *
  * This software is licensed as described in the file COPYING, which
  * you should have received as part of this distribution. The terms
@@ -17,6 +17,8 @@
  *
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
+ *
+ * SPDX-License-Identifier: curl
  *
  ***************************************************************************/
 #include "curlcheck.h"
@@ -52,13 +54,13 @@ UNITTEST_START
   for(i = 0; i < sizeof(tests)/sizeof(tests[0]); i++) {
     timediff_t result = Curl_timediff(tests[i].first, tests[i].second);
     if(result != tests[i].result) {
-      printf("%d.%06u to %d.%06u got %d, but expected %d\n",
-             tests[i].first.tv_sec,
+      printf("%ld.%06u to %ld.%06u got %d, but expected %ld\n",
+             (long)tests[i].first.tv_sec,
              tests[i].first.tv_usec,
-             tests[i].second.tv_sec,
+             (long)tests[i].second.tv_sec,
              tests[i].second.tv_usec,
-             result,
-             tests[i].result);
+             (int)result,
+             (long)tests[i].result);
       fail("unexpected result!");
     }
   }
